Make add and remove abstract & fix bug when emitting initial state
