+++ src/main/java/jtorrent/domain/util/RxObservableCollection.java
@@ -2,6 +2,8 @@ package jtorrent.domain.util;
 
 import java.util.Collection;
 
+import io.reactivex.rxjava3.core.Observer;
+
 public abstract class RxObservableCollection<E>
         extends RxObservableCollectionBase<E, Collection<E>, CollectionEvent<E>> {
 
@@ -10,13 +12,26 @@ public abstract class RxObservableCollection<E>
     }
 
     @Override
-    protected void notifyAdded(E item) {
-        emitEvent(CollectionEvent.add(item));
+    protected void add(E item) {
+        if (collection.add(item)) {
+            notifyAdded(item);
+        }
     }
 
     @Override
-    protected void notifyRemoved(E item) {
+    protected void remove(E item) {
+        if (collection.remove(item)) {
+            notifyRemoved(item);
+        }
+    }
+
+    private void notifyRemoved(E item) {
         emitEvent(CollectionEvent.remove(item));
+    }
+
+    @Override
+    protected void emitInitialState(Observer<? super CollectionEvent<E>> observer) {
+        collection.forEach(item -> observer.onNext(CollectionEvent.add(item)));
     }
 
     @Override
@@ -24,8 +39,7 @@ public abstract class RxObservableCollection<E>
         emitEvent(CollectionEvent.clear());
     }
 
-    @Override
-    protected void emitInitialState() {
-        collection.forEach(this::notifyAdded);
+    private void notifyAdded(E item) {
+        emitEvent(CollectionEvent.add(item));
     }
 }
